Now Bagging (and RandomForest) can process datasets with more than 22 million instances. This one was very mysterious at first.
git-svn-id: https://svn.cms.waikato.ac.nz/svn/weka/trunk@11465 e0a1b77d-ad91-4216-81b1-defd5f83fa92

+++ weka/src/main/java/weka/classifiers/meta/Bagging.java
@@ -504,7 +504,7 @@ public class Bagging
    */
   @Override
   protected synchronized Instances getTrainingSet(int iteration) throws Exception {
-    int bagSize = m_data.numInstances() * m_BagSizePercent / 100;
+    int bagSize = (int) (m_data.numInstances() * (m_BagSizePercent / 100.0));
     Instances bagData = null;
     Random r = new Random(m_Seed + iteration);
